Stop monkey patching the matcher class.
The monkey patching code originated in rspec-rails
when it was a separate code base from the matcher
definition in rspec-expectations, but now that they
live in the same code base, it's best not to
monkey patch.

Also, we don't need to limit this to just when rails
is loaded; nothing in the logic relies on rails. We
were only relying on rails for `alias_method_chain` and
`ActiveSupport::Concern`.